Adding Confirm Dialogue before Deleting Todos

+++ app.js
@@ -1,6 +1,7 @@
 const addTodoBtn = document.getElementById("addTodoBtn")
 const inputTag = document.getElementById("todoInput")
 const todoListUl = document.getElementById("todoList")
+const remaining = document.getElementById("remaining-count")
 
 let todoText; // This should be populated when the user clicks on Add button
 let todos = [];
@@ -10,16 +11,19 @@ if (todosString) {
     todos = JSON.parse(todosString);
 }
 
+
+
 const populateTodos = () => {
-    let string = "";
+    let string = ""; 
     for (const todo of todos) {
-        string += `<li id="todo-${todo.id}" class="todo-item ${todo.isCompleted ? "completed" : ""}">
+        string += `<li id="${todo.id}" class="todo-item ${todo.isCompleted ? "completed" : ""}">
             <input type="checkbox" class="todo-checkbox" ${todo.isCompleted ? "checked" : ""} >
             <span class="todo-text">${todo.title}</span>
             <button class="delete-btn">×</button>
-        </li>`
+        </li>` 
     }
     todoListUl.innerHTML = string
+
 
     // Add the checkbox logic to populate todos
     const todoCheckboxes = document.querySelectorAll(".todo-checkbox")
@@ -28,53 +32,65 @@ const populateTodos = () => {
         element.addEventListener("click", (e) => {
             if (e.target.checked) {
                 element.parentNode.classList.add("completed")
+                console.log(todos)
                 // Grab this todo from todos array and update the todos array to set this todo's isCompleted attribute as true
                 todos = todos.map(todo => {
-                    if ("todo-" + todo.id == element.parentNode.id) {
+                    if (todo.id == element.parentNode.id) {
+                        console.log(todo.id, element.parentNode.id)
                         return { ...todo, isCompleted: true }
                     }
                     else {
                         return todo
                     }
                 })
+                remaining.innerHTML = todos.filter((item)=>{return item.isCompleted!=true}).length;
                 localStorage.setItem("todos", JSON.stringify(todos))
             }
             else {
                 element.parentNode.classList.remove("completed")
                 // Grab this todo from todos array and update the todos array to set this todos isCompleted attribute as false
                 todos = todos.map(todo => {
-                    if ("todo-" + todo.id == element.parentNode.id) {
+                    if (todo.id == element.parentNode.id) {
                         return { ...todo, isCompleted: false }
                     }
                     else {
                         return todo
                     }
                 })
+                remaining.innerHTML = todos.filter((item)=>{return item.isCompleted!=true}).length;
                 localStorage.setItem("todos", JSON.stringify(todos))
             }
         })
     })
+
+
+
+
 
     // Handle the delete buttons
     let deleteBtns = document.querySelectorAll(".delete-btn")
 
     deleteBtns.forEach((element) => {
         element.addEventListener("click", (e) => {
-            todos = todos.filter((todo) => {
-                return ("todo-" + todo.id) !== (e.target.parentNode.id)
-            })
-            localStorage.setItem("todos", JSON.stringify(todos))
-            populateTodos()
+            const confirmation = confirm("Do you want to delete this todo") 
+            if(confirmation){ 
+                todos = todos.filter((todo) => {
+                    return (todo.id) !== (e.target.parentNode.id)
+                })
+                localStorage.setItem("todos", JSON.stringify(todos))
+                populateTodos()
+            }
         })
     })
-
+    
 }
+
 
 addTodoBtn.addEventListener("click", () => {
     todoText = inputTag.value
-    //Check if the length of todo is greater than 3
+    // check if the length of todo is greater than 3 
     if(todoText.trim().length<4){
-        alert("You cannot add todo that small!")
+        alert("You cannot add a todo that small!")
         return
     }
     inputTag.value = ""
